# aperag/api/openapi.merged.yaml
openapi: 3.1.0
info:
  title: ApeRAG API
  description: ApeRAG API Documentation
  version: 1.0.0
servers:
  - url: /api/v1
    description: API v1
paths:
...
  /web/search:
    post:
      summary: Web search
      description: Perform web search using various search engines
      tags:
        - web
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/webSearchRequest'
            examples:
              simple_search:
                summary: Simple search
                description: Basic web search
                value:
                  query: ApeRAG 2025年最新发展
                  max_results: 5
              advanced_search:
                summary: Advanced search
                description: Search with custom parameters
                value:
                  query: 机器学习最新进展
                  max_results: 10
                  timeout: 30
                  locale: zh-CN
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/webSearchResponse'
              examples:
                search_results:
                  summary: Search results
                  description: Successful search with results
                  value:
                    query: ApeRAG 2025年最新发展
                    results:
                      - rank: 1
                        title: ApeRAG 2025年技术路线图
                        url: https://example.com/aperag-2025-roadmap
                        snippet: ApeRAG在2025年将重点发展...
                        domain: example.com
                        timestamp: '2025-01-01T00:00:00Z'
                    total_results: 1
                    search_time: 1.2
        '400':
          description: Bad request - Invalid search parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
              examples:
                empty_query:
                  summary: Empty query
                  description: Query cannot be empty
                  value:
                    code: '400'
                    message: Search query cannot be empty
                invalid_engine:
                  summary: Invalid search engine
                  description: Unsupported search engine
                  value:
                    code: '400'
                    message: 'Unsupported search engine: invalid_engine'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          description: Internal server error - Search failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
              examples:
                search_error:
                  summary: Search failed
                  description: Search operation failed
                  value:
                    code: '500'
                    message: 'Search failed: Network timeout'
  /web/read:
    post:
      summary: Web content reading
      description: Read and extract content from web pages
      tags:
        - web
      security:
        - BearerAuth: []
        - CookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/webReadRequest'
            examples:
              single_url:
                summary: Single URL
                description: Read content from a single URL
                value:
                  urls: https://example.com/article
                  timeout: 30
              multiple_urls:
                summary: Multiple URLs
                description: Read content from multiple URLs
                value:
                  urls:
                    - https://example.com/article1
                    - https://example.com/article2
                  max_concurrent: 2
              advanced_reading:
                summary: Advanced content extraction
                description: Content extraction with CSS selectors
                value:
                  urls: https://example.com/spa-page
                  css_selector: article .content
                  wait_for_selector: .loading-complete
                  exclude_selector: .ads, .sidebar
                  timeout: 45
                  locale: zh-CN
      responses:
        '200':
          description: Content reading completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/webReadResponse'
              examples:
                single_success:
                  summary: Single URL success
                  description: Successfully read single URL
                  value:
                    results:
                      - url: https://example.com/article
                        status: success
                        title: 示例文章标题
                        content: |-
                          # 示例文章

                          这是文章内容...
                        extracted_at: '2025-01-07T10:00:00Z'
                        word_count: 500
                        token_count: 750
                    total_urls: 1
                    successful: 1
                    failed: 0
                    processing_time: 2.5
                mixed_results:
                  summary: Mixed results
                  description: Some URLs succeeded, some failed
                  value:
                    results:
                      - url: https://example.com/article1
                        status: success
                        title: 成功的文章
                        content: |-
                          # 成功读取

                          内容...
                        extracted_at: '2025-01-07T10:00:00Z'
                        word_count: 300
                        token_count: 450
                      - url: https://invalid-url.com/404
                        status: error
                        error: 页面无法访问
                        error_code: TIMEOUT
                    total_urls: 2
                    successful: 1
                    failed: 1
                    processing_time: 5.2
        '400':
          description: Bad request - Invalid read parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
              examples:
                empty_urls:
                  summary: Empty URLs
                  description: URLs list cannot be empty
                  value:
                    code: '400'
                    message: URLs list cannot be empty
                invalid_url:
                  summary: Invalid URL format
                  description: URL format is invalid
                  value:
                    code: '400'
                    message: Invalid URL format
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
        '500':
          description: Internal server error - Reading failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/failResponse'
              examples:
                read_error:
                  summary: Reading failed
                  description: Content reading operation failed
                  value:
                    code: '500'
                    message: 'Content reading failed: Provider error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
    CookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
  schemas:
...
    failResponse:
      type: object
      properties:
        code:
          type: string
          description: Error code
          example: '400'
        message:
          type: string
          description: Error message
          example: Invalid request
...
    webSearchRequest:
      type: object
      description: Web search request
      properties:
        query:
          type: string
          description: Search query for regular web search. Optional if only using LLM.txt discovery.
          example: ApeRAG 2025年最新发展
        max_results:
          type: integer
          description: Maximum number of results to return
          default: 5
          example: 5
        timeout:
          type: integer
          description: Request timeout in seconds
          default: 30
          example: 30
        locale:
          type: string
          description: Browser locale
          default: en-US
          example: en-US
        source:
          type: string
          description: Domain or URL for site-specific filtering. When provided with query, limits search results to this domain (e.g., 'site:vercel.com query').
          example: vercel.com
        search_llms_txt:
          type: string
          description: Domain for LLM.txt discovery search. When provided, performs additional LLM-optimized content discovery from the specified domain, independent of the main search. Results are merged with regular search results.
          example: anthropic.com
      required: []
    webSearchResultItem:
      type: object
      description: Individual web search result
      properties:
        rank:
          type: integer
          description: Result rank
          example: 1
        title:
          type: string
          description: Page title
          example: ApeRAG 2025年技术路线图
        url:
          type: string
          description: Page URL
          example: https://example.com/aperag-2025-roadmap
        snippet:
          type: string
          description: Page snippet
          example: ApeRAG在2025年将重点发展...
        domain:
          type: string
          description: Domain name
          example: example.com
        timestamp:
          type: string
          format: date-time
          description: Result timestamp
          example: '2025-01-01T00:00:00Z'
      required:
        - rank
        - title
        - url
        - snippet
        - domain
    webSearchResponse:
      type: object
      description: Web search response
      properties:
        query:
          type: string
          description: Original search query
        results:
          type: array
          description: List of search results
          items:
            $ref: '#/components/schemas/webSearchResultItem'
        total_results:
          type: integer
          description: Total number of results found
        search_time:
          type: number
          format: float
          description: Search time in seconds
      required:
        - query
        - results
    webReadRequest:
      type: object
      description: Web content reading request
      properties:
        url_list:
          type: array
          description: List of URLs to read (for single URL, use array with one element)
          items:
            type: string
          example:
            - https://example.com/article
        timeout:
          type: integer
          description: Request timeout in seconds
          default: 30
          example: 30
        locale:
          type: string
          description: Browser locale
          default: en-US
          example: en-US
        max_concurrent:
          type: integer
          description: Maximum concurrent requests for multiple URLs
          default: 3
          example: 3
        method:
          type: string
          enum:
            - GET
            - POST
            - PUT
            - PATCH
            - DELETE
          description: HTTP method to use. POST/PUT/PATCH are useful for JS-rendered pages whose data comes from AJAX endpoints (e.g. search result APIs)
          default: GET
          example: POST
        body:
          type: object
          description: Request body for POST/PUT/PATCH methods (JSON object, keys and values are sent as-is)
          default: {}
          example:
            key: search keyword
            page: 1
        body_type:
          type: string
          enum:
            - form
            - json
          description: How to encode the body — "form" sends application/x-www-form-urlencoded, "json" sends application/json
          default: form
          example: form
        extra_headers:
          type: object
          description: Additional HTTP headers to send with the request (e.g. Referer, X-Requested-With, Cookie). Merged over the default headers.
          default: {}
          example:
            Referer: https://www.scm.com.cn/search
            X-Requested-With: XMLHttpRequest
      required:
        - url_list
    webReadResultItem:
      type: object
      description: Individual web content reading result
      properties:
        url:
          type: string
          description: Requested URL
        status:
          type: string
          enum:
            - success
            - error
          description: Processing status
        title:
          type: string
          description: Page title
        content:
          type: string
          description: Extracted content in Markdown format
        extracted_at:
          type: string
          format: date-time
          description: Content extraction timestamp
        word_count:
          type: integer
          description: Word count of content
        token_count:
          type: integer
          description: Estimated token count
        error:
          type: string
          description: Error message if failed
        error_code:
          type: string
          description: Error code if failed
      required:
        - url
        - status
    webReadResponse:
      type: object
      description: Web content reading response
      properties:
        results:
          type: array
          description: List of reading results
          items:
            $ref: '#/components/schemas/webReadResultItem'
        total_urls:
          type: integer
          description: Total number of URLs processed
        successful:
          type: integer
          description: Number of successful extractions
        failed:
          type: integer
          description: Number of failed extractions
        processing_time:
          type: number
          format: float
          description: Total processing time in seconds
      required:
        - results
        - total_urls
        - successful
        - failed
...